[refactor] : split up component for better readability
split up component for better readability
