+++ challenge-react/app/error.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import React from "react";
+
+export default function error() {
+  return (
+    <div className=" text-sm text-balance flex-1 flex-col  col-span-1 md:col-span-2 lg:col-span-3 text-center flex justify-center items-center  w-full h-full">
+      <p>Ouch, Something went wrong :(</p>
+    </div>
+  );
+}
